refact: userCard 외부 파일로 분리 #154
